prevented error for scheduler not started

+++ minimal-scheduler/stop.sh
@@ -1,4 +1,6 @@
 #!/bin/sh
 
 # Remove the scheduler
-sudo rm /sys/fs/bpf/sched_ext/sched_ops
+if [ -e /sys/fs/bpf/sched_ext/sched_ops ]; then
+    sudo rm /sys/fs/bpf/sched_ext/sched_ops
+fi
